added commands for retrieveing versions
Added information on how to retrieve component versions in 'system configuration', because this is where the documentation first begins using commands in BISDN Linux OS.
Added the commands 'opkg info', 'less /etc/os-release' for information on services/packages and BISDN Linux.
Additionally added 'uname -a', as it gives the Kernel version and build date which might be useful.