Updated and restructured the page on getting started.
Added sudo to all commands wich needed it.
Removed double referenece to opkg for checking package versions
Moved check for services that are enabled/disabled to a section right after explanation on how to check services. (i.e. sudo systemctl status), so it's more understandable HOW to check if a service is enabled.