Fix internal links
This changeset fixes broken internal links in the website.

Absolute paths pointing to local Markdown files don't work because
jekyll cannot find the files and leaves the paths as is.

For instance, having this in getting_started/basic_networking.md does
not work:

  [Network Configuration](/network_configuration.md)

If we use relative links instead, jekyll finds the files and rewrites the
reference from .md to .html which results in working URLs:

  [Network Configuration](../network_configuration.md)