stp.md: add command to set bridge up
Currently the users might run into issues if they delay setting an stp bridge
up until after adding ports to the bridge. If we add this step to our docs it
might save users from running into this problem.
Since the docs then go on to describe 'configuring and attaching the ports to
the bridge', it fits with the flow of the documentation that the bridge's state
should be UP.

Signed-off-by: Hilmar Magnusson <hilmar.magnusson@bisdn.de>